🧪 Move simple example in test directory
It's now runnable from Intellij
